GraphicsEngine_v1_01
This update marks the second major upgrade of the 3D graphics engine. The user can now create a user defined console window and display a cube.

The cube has to be initialized in the EngineCreate() function. The user can call the demoCube() function inside the EngineUpdate() function. While the console window is active, a cube will be drawn and rotated.