Added Object3D Class
Like the Cube class, the Object3D class inherits from the Polygon3D class. It creates a mesh from the data of a .obj file. It doesn't work quite yet, because the triangles are being drawn randomly but they should be drawn in clockwise order. This should be implemented in the next update.